[OGUI-867] Refactor and add label for AliECS loading
